Added new data, organized file tree, minor bug fixes
Found new data, but there's one issue: there's not enough.

**Changes**
- Added data 2014-2017, both Google queries and CDC data. *Google queries are now organized in one CSV file
- Reworked file tree. Added region and year specific folders for data.

There's one issue with the data I have: there's not enough, and although the data I need exists, it's not in an easily accessible format. Instead, it's in a .csv file for every week of the year. I plan to build a simple web scraper to address this.

God, sometimes I sound like HR and sometimes I sound like someone sent up to HR.