Potentially fixed unusually high loss, cleaned file tree, removed Herobrine
It works!
**Changes**
- Split x_arr and y_arr into training/testing arrays, 80/20 split.
- Plotted model predictions.
- Cleaned file tree and removed/renamed redundant files (preprocessingfinal >> preprocessing2, moved to /oldscripts)
- Added RMSE evaluation after model fitting. Currently an impromptu sklearn call, will implement with Keras later.

It turns out that nothing was wrong with the network itself. Evaluating the model with RMSE as a metric for accuracy returns a mean square error of 5. Considering that outbreaks of E. coli usually result in upwards of 10 cases, I think this model is a success. I haven't even gathered the whole data set yet! I would be satisfied with taking this to the science fair.